requirements.txt
streamlit
pdf2docx
Pillow
pytesseract
pdf2image
python-docx